LPD-72635 Move methods to the existing interface and also move the implementation
